updated Bill Report

+++ Details/Show.hpp
@@ -8,14 +8,12 @@ public:
       if (C.is_open() && TestofPostgres() == 1) {
         std::cout << "Opened database successfully: " << C.dbname() << '\n';
       }
-      // PostgresConnection connection;
-      // connection.Conn();
       nontransaction N(C);
       result R( N.exec( sql ));
       for (result::const_iterator c = R.begin(); c != R.end(); c++) {
         cout << "PRODUCTNAME = " << c[1].as<string>() << endl;
         cout << "BRANDNAME = " << c[2].as<string>() << endl;
-        cout << "PRICE = " << c[3].as<int>() << endl;
+        cout << "PRICE = " << c[3].as<double>() << endl;
         cout << "PRODUCTYEAR = " << c[4].as<int>() << endl;
       }
       cout << "Operation done successfully" << endl;
